Implement FIB table based on hash table algorithm from Won et al 2013
Implement FIB functionality based on the algorithm described in "Named data
networking on a router: Fast and DoS-resistant forwarding with hash tables"
by W. So, A. Narayanan and D. Oran, published in Architectures for Networking
and Communications Systems, 2013.

The idea is to create virtual prefixes with length M and then do a
2-stage longest prefix match (LPM) using hash tables. For an incoming name,
if its length is less than M, do standard LPM. If its length is greater
than or equal to M, look up the virtual name of length M, which is a
name with length M that is a prefix of the input name. This virtual name
contains an attribute called max depth (MD) that describes the longest
name associated with this virtual name. Do LPM starting with a name of
length MD. This ensures that regardless of how long the input name is,
we will do no more than O(MD) lookups, where MD is the longest name.